+++ include/utils/ColorBgr.h
@@ -49,19 +49,26 @@ struct ColorBgr
 	{
 	}
 
-	ColorBgr(ColorRgb rgb):
-			red(rgb.red),
+	explicit ColorBgr(ColorRgb rgb):
+			blue(rgb.blue),
 			green(rgb.green),
-			blue(rgb.blue)
+			red(rgb.red)
 	{
 	}
 
-	ColorBgr operator-(const ColorBgr& b) const
+    ColorBgr& operator=(const ColorRgb& rhs)
+    {
+        blue = rhs.blue;
+        green = rhs.green;
+        red = rhs.red;
+        return *this;
+    }
+
+	friend inline ColorBgr operator-(ColorBgr a, const ColorBgr &b)
 	{
-		ColorBgr a(*this);
-		a.red -= b.red;
+		a.blue -= b.blue;		
 		a.green -= b.green;
-		a.blue -= b.blue;
+		a.red -= b.red;
 		return a;
 	}