+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🔧 Changed
 
 - **Fixes:**
+  - Image Effect trigger kills Hyperion (#1980)
 
 ---
 
@@ -26,6 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Technical
 
+- Fixed - Locally installed flatc/protoc compilers were not found
+- ProviderRestAPI - Handle QNetworkAccessManager as direct object 
 
 ## [2.2.0](https://github.com/hyperion-project/hyperion.ng/releases/tag/2.2.0) - 2026-02-02